requirements.txt
# Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Database
supabase>=2.3.0
python-dotenv>=1.0.0

# API (for later phases)
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Utilities
loguru>=0.7.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
httpx>=0.26.0
